+++ LibCarla/source/carla/geom/Vector2D.h
@@ -39,6 +39,10 @@ namespace geom {
     // -- Other methods --------------------------------------------------------
     // =========================================================================
 
+    double ExactLength() const {
+       return std::sqrt(double(x) * x + double(y) * y);
+    }
+
     float SquaredLength() const {
       return x * x + y * y;
     }
@@ -47,11 +51,24 @@ namespace geom {
        return std::sqrt(SquaredLength());
     }
 
-    Vector2D MakeUnitVector() const {
-      const float len = Length();
-      DEVELOPMENT_ASSERT(len > 2.0f * std::numeric_limits<float>::epsilon());
-      const float k = 1.0f / len;
-      return Vector2D(x * k, y * k);
+    Vector2D MakeUnitVectorLengthInput(const double length, const float epsilon = 2.0f * std::numeric_limits<float>::epsilon()) const {
+      if (length < epsilon) {
+        return *this;
+      }
+      const double k = 1.0 / length;
+      Vector2D result(float(x * k), float(y * k));
+      return result;
+    }
+
+    Vector2D MakeUnitVector(const float epsilon = 2.0f * std::numeric_limits<float>::epsilon()) const  {
+      const double length = ExactLength();
+      return MakeUnitVectorLengthInput(length, epsilon);
+    }
+
+    std::pair<Vector2D, double> GetUnitVectorAndLength() const {
+      const auto length = ExactLength();
+      const auto unit_vector = MakeUnitVectorLengthInput(length);
+      return std::make_pair(unit_vector, length);
     }
 
     // =========================================================================