+++ PythonAPI/test/unit/test_transform.py
@@ -147,7 +147,8 @@ class TestTransform(unittest.TestCase):
         point = carla.Location(x=0.0, y=0.0, z=2.0)
         t.transform(point)
 
-        self.assertTrue(abs(point.x - (-2.0)) <= error)
+        #!!! This solution_list was wrong in CARLA version <= 0.9.16 due to invalid pitch AND roll rotations, (most relevant) yaw rotations were not affected
+        self.assertTrue(abs(point.x - (2.0)) <= error)
         self.assertTrue(abs(point.y - 0.0) <= error)
         self.assertTrue(abs(point.z - (-1.0)) <= error)
 
@@ -163,9 +164,10 @@ class TestTransform(unittest.TestCase):
                       ]
         t.transform(point_list)
 
-        solution_list = [carla.Location(-2.0, 0.0, -1.0),
-                         carla.Location(-1.0, 10.0, -1.0),
-                         carla.Location(-2.0, 18.0, -1.0)
+        #!!! This solution_list was wrong in CARLA version <= 0.9.16 due to invalid pitch AND roll rotations, (most relevant) yaw rotations were not affected
+        solution_list = [carla.Location(2.0, 0.0, -1.0),   
+                         carla.Location(1.0, 10.0, -1.0),
+                         carla.Location(2.0, 18.0, -1.0)
                          ]
 
         for i in range(len(point_list)):
@@ -185,9 +187,10 @@ class TestTransform(unittest.TestCase):
                       ]
         t.transform(point_list)
 
-        solution_list = [carla.Vector3D(-2.0, 0.0, -1.0),
-                         carla.Vector3D(-1.0, 10.0, -1.0),
-                         carla.Vector3D(-2.0, 18.0, -1.0)
+        #!!! This solution_list was wrong in CARLA version <= 0.9.15 due to invalid pitch AND roll rotations, (most relevant) yaw rotations were not affected
+        solution_list = [carla.Vector3D(2.0, 0.0, -1.0),
+                         carla.Vector3D(1.0, 10.0, -1.0),
+                         carla.Vector3D(2.0, 18.0, -1.0)
                          ]
 
         for i in range(len(point_list)):