+++ CHANGELOG.md
@@ -4,6 +4,8 @@
  * Improved the turning behavior of vehicles controlled by the TrafficManager, making them smoother.
  * Added a hybrid solid-state LiDAR with adjustable parameters (blueprint attributes)
  * Fix OpenDrive Builder lane width
+ * Introduced geom::AngularVelocity, geom::Velocity, geom::Acceleration, geom::Quaternion types
+ * Fixed geom::Rotation::RotateVector() rotation directions of pitch and roll
 
 ## CARLA 0.9.16